requirements.txt
Flask==1.1.1
redis==3.3.11
numpy==1.18.1
pandas==0.25.3
PyYaml==5.3
sqlalchemy==1.3.12
mysql-connector-python==8.0.19